# spring-server/src/main/resources/api-spec.yaml
openapi: 3.0.1
info:
  title: Articles app API 
  description: "API de mon grand projet qui n'a pas encore de nom. App sur laquelle on peut poster des articles et intéragir avec ceux-ci."
  version: 1.0.0
servers:
- url: https://localhost:8080/api
...
paths:
...
  /categories:
    get:
      tags:
        - categories
      summary: Finds all categories
      operationId: findCategories
      responses: 
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Category'
              
    post:
      tags:
        - categories
      summary: Add a new category
      operationId: addCategory
      requestBody:
        description: Category that needs to be added
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Category'
        required: true
      responses:
        200: 
          description: successful operation
        405:
          description: Invalid input
          content: {}
  /categories/{categoryId}:
    get:
      tags:
        - categories
      summary: finds a category by Id
      operationId: findACategoryById
      parameters: 
        - name: categoryId
          in: path
          description: category to delete
          required: true
          schema:
            type: string
      responses:
        200:
          description: successful operation
    patch:
      tags:
        - categories
      summary: updates a category by Id
      operationId: updateCategoryById
      parameters: 
        - name: categoryId
          in: path
          description: category to delete
          required: true
          schema:
            type: string
      requestBody:
        description: updated category
        content:
          '*/*':
            schema:
              $ref: '#/components/schemas/Category'
        required: true
      responses:
        200:
          description: successful operation
    delete:
        tags: 
          - categories
        parameters: 
        - name: categoryId
          in: path
          description: category to delete
          required: true
          schema:
            type: string
        summary: deletes a category by ID
        operationId: deleteACategory
        responses:
          200:
            description: successful operation
  /tags:
    get:
      tags:
        - tags
      summary: Finds all tags
      operationId: findtags
      responses: 
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Tag'
              
    post:
      tags:
        - tags
      summary: Add a new tag
      operationId: addTag
      requestBody:
        description: Tag that needs to be added
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Tag'
        required: true
      responses:
        200: 
          description: successful operation
        405:
          description: Invalid input
          content: {}
  /tags/{tagId}:
    get:
      tags:
        - tags
      summary: finds a tag by Id
      operationId: findATagById
      parameters: 
        - name: tagId
          in: path
          description: Tag to delete
          required: true
          schema:
            type: string
      responses:
        200:
          description: successful operation
    patch:
      tags:
        - tags
      summary: updates a tag by Id
      operationId: updateTagById
      parameters: 
        - name: tagId
          in: path
          description: tag to delete
          required: true
          schema:
            type: string
      requestBody:
        description: updated tag
        content:
          '*/*':
            schema:
              $ref: '#/components/schemas/Tag'
        required: true
      responses:
        200:
          description: successful operation
    delete:
        tags: 
          - tags
        parameters: 
        - name: tagId
          in: path
          description: tag to delete
          required: true
          schema:
            type: string
        summary: deletes a tag by ID
        operationId: deleteAtag
        responses:
          200:
            description: successful operation
components:
  schemas:
...
    Tag:
      required:
        - id 
        - name
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
...
    Category:
      required:
        - name
        - children
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        children:
          type: array
          items:
            $ref: '#/components/schemas/Tag'